requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-doc==0.0.4
annotated-types==0.6.0
anyio==4.12.1
attrs==23.2.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.3.1
colorama==0.4.6
dataclasses-json==0.6.5
distro==1.9.0
fastapi==0.129.0
frozenlist==1.4.1
greenlet==3.0.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.7
jiter==0.13.0
jsonpatch==1.33
jsonpointer==2.4
langchain==0.1.17
langchain-community==0.0.36
langchain-core==0.1.48
langchain-text-splitters==0.0.1
langsmith==0.1.52
marshmallow==3.21.1
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==2.20.0
orjson==3.10.2
packaging==23.2
pydantic==2.7.1
pydantic_core==2.18.2
python-dotenv==1.2.1
PyYAML==6.0.1
requests==2.31.0
sniffio==1.3.1
SQLAlchemy==2.0.29
starlette==0.52.1
tenacity==8.2.3
tqdm==4.67.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.2.1
uvicorn==0.40.0
yarl==1.9.4
